+++ src/lib/components/metrics/browserDistribution.svelte
@@ -2,6 +2,7 @@
 	import type { Usage } from '@prisma/client';
 
 	import { getLocalTimeZone } from '@internationalized/date';
+	import { browser } from '$app/environment';
 	import { page } from '$app/state';
 	import { getTranslations } from '$lib/i18n/index.svelte';
 	import { getMetricsStore } from '$lib/stores/metrics.svelte';
@@ -13,6 +14,8 @@
 
 	let data = $state<{ id: null | string; name: null | string; value: number }[]>([]);
 	const loadData = async () => {
+		if(!browser) return
+
 		try {
 			const res = (await (
 				await (page.data.fetch as typeof fetch)(