+++ src/lib/components/settings/adminPanel/component.svelte
@@ -2,20 +2,25 @@
 	import { applyAction, enhance } from '$app/forms';
 	import { invalidateAll } from '$app/navigation';
 	import P from '$lib/components/typography/text/p.svelte';
+	import { Input } from '$lib/components/ui/input';
 	import { Label } from '$lib/components/ui/label';
 	import Separator from '$lib/components/ui/separator/separator.svelte';
 	import Switch from '$lib/components/ui/switch/switch.svelte';
 	import { getTranslations } from '$lib/i18n/index.svelte';
+	import { prefersReducedMotion } from 'svelte/motion';
+	import { fly } from 'svelte/transition';
 	const i18n = getTranslations();
 
 	let {
 		allowUnsecureHTTP,
+		customRedirect=$bindable(),
 		disableHome,
 		enabledMFA,
 		enableLimits,
-		enableSignup
+		enableSignup,
 	}: {
 		allowUnsecureHTTP: boolean;
+		customRedirect:null|string,
 		disableHome: boolean;
 		enabledMFA: boolean;
 		enableLimits: boolean;
@@ -54,6 +59,8 @@
 			value: enabledMFA === true || false
 		}
 	]);
+
+
 </script>
 
 {#each items as item}
@@ -76,6 +83,23 @@
 			/>
 		</div>
 	</div>
+	{#if item.id ==="DisableHome" && disableHome === true}
+	<form class="w-full flex flex-col" transition:fly|global={{delay: prefersReducedMotion.current ? 0: 400, y: prefersReducedMotion.current ? 0:12}}	
+	method="post"
+	id="customRedirect"
+	action="?/customRedirect"
+	use:enhance={({ formData }) => {
+		if(customRedirect) formData.set('customRedirect',customRedirect );
+		return async ({ result }) => {
+			await applyAction(result);
+			await invalidateAll();
+		};
+	}}><Label class="pb-2">{i18n.t('settings.label.custom-home-redirect')}</Label>
+	<Input name="customRedirect" onchange={async()=>{
+		document.forms.namedItem('customRedirect')?.requestSubmit()
+	}} type="text" bind:value={customRedirect} />
+	</form>
+	{/if}
 	<Separator />
 {/each}