+++ CHANGELOG.md
@@ -272,3 +272,7 @@
   - Restored URL creation via adding them to groups only [read discussion...](https://github.com/urania-dev/snapp/discussions/67#discussioncomment-12656742)
 - 0.9-rc-014
   - Fixed: [#107 Random Error At Night](https://github.com/urania-dev/snapp/issues/107) now OIDC get parsed on demand and not only on app startup
+- 0.9-rc-015-18
+  - Fixed (i hope): [#113 Errors while using it in Production](https://github.com/urania-dev/snapp/issues/113) fixed error in metrics pages due to misrendering across serverside and client side
+  - [#112 sqlite file not copied](https://github.com/urania-dev/snapp/issues/112) entrypoint.sh should now check if sqlite db exists or create it
+  - [#111 Default page redirection](https://github.com/urania-dev/snapp/issues/111) now added a field in settings for this